Move SearchResultMapper to search package to conform to project structure
