+++ src/test/java/de/uol/pgdoener/civicsage/mapper/SearchResultMapperTest.java
@@ -2,6 +2,7 @@ package de.uol.pgdoener.civicsage.mapper;
 
 
 import de.uol.pgdoener.civicsage.business.dto.SearchResultDto;
+import de.uol.pgdoener.civicsage.search.SearchResultMapper;
 import org.junit.jupiter.api.Test;
 import org.springframework.ai.document.Document;
